requirements.txt
python-dotenv
requests
rdflib
xlrd
pyyaml

